# spec/api/openapi.yaml
openapi: 3.2.1
...
info:
  title: Knowledge Platform Document API
  version: 0.1.0
  description: One authorized Document API for human, agent and CLI clients. Identity is established only by a trusted
    adapter; request data never supplies the actor.
servers:
- url: /
  description: Same-origin deployment; no production endpoint is fixed.
...
paths:
...
  /v1/documents/{documentId}/versions:
    parameters:
    - $ref: '#/components/parameters/documentId'
    get:
      tags:
      - Versions
      summary: List authorized versions
      operationId: listDocumentVersions
      description: The trusted identity adapter supplies the verified actor. Authorization and Audit remain in Application.
      responses:
        '200':
          $ref: '#/components/responses/VersionList'
        '401':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
      parameters:
      - $ref: '#/components/parameters/purpose'
      - $ref: '#/components/parameters/pageSize'
      - $ref: '#/components/parameters/cursor'
    post:
      tags:
      - Versions
      summary: Create a working version
      operationId: createDocumentVersion
      description: The trusted identity adapter supplies the verified actor. Authorization and Audit remain in Application.
      responses:
        '201':
          $ref: '#/components/responses/VersionMutationResult'
        '401':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
      requestBody:
        $ref: '#/components/requestBodies/VersionMultipart'
  /v1/documents/{documentId}/versions/{versionId}:
    parameters:
    - $ref: '#/components/parameters/documentId'
    - $ref: '#/components/parameters/versionId'
    get:
      tags:
      - Versions
      summary: Read an authorized version
      operationId: getDocumentVersion
      description: The trusted identity adapter supplies the verified actor. Authorization and Audit remain in Application.
      responses:
        '200':
          $ref: '#/components/responses/Version'
        '401':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
      parameters:
      - $ref: '#/components/parameters/purpose'
    put:
      tags:
      - Versions
      summary: Update a working version
      operationId: updateWorkingVersion
      description: The trusted identity adapter supplies the verified actor. Authorization and Audit remain in Application.
      responses:
        '200':
          $ref: '#/components/responses/VersionMutationResult'
        '401':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
      requestBody:
        $ref: '#/components/requestBodies/VersionMultipart'
  /v1/documents/{documentId}/versions/{versionId}:rebase:
    parameters:
    - $ref: '#/components/parameters/documentId'
    - $ref: '#/components/parameters/versionId'
    post:
      tags:
      - Versions
      summary: Rebase a working version
      operationId: rebaseWorkingVersion
      description: The trusted identity adapter supplies the verified actor. Authorization and Audit remain in Application.
      responses:
        '200':
          $ref: '#/components/responses/VersionMutationResult'
        '401':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
      requestBody:
        $ref: '#/components/requestBodies/RebaseVersion'
  /v1/documents/{documentId}/versions/{versionId}:publish:
    parameters:
    - $ref: '#/components/parameters/documentId'
    - $ref: '#/components/parameters/versionId'
    post:
      tags:
      - Lifecycle
      summary: Publish a version
      operationId: publishVersion
      description: The trusted identity adapter supplies the verified actor. Authorization and Audit remain in Application.
      responses:
        '200':
          $ref: '#/components/responses/PublishResult'
        '401':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
      requestBody:
        $ref: '#/components/requestBodies/PublishVersion'
  /v1/documents/{documentId}/versions/{versionId}:withdraw:
    parameters:
    - $ref: '#/components/parameters/documentId'
    - $ref: '#/components/parameters/versionId'
    post:
      tags:
      - Lifecycle
      summary: Withdraw a version
      operationId: withdrawVersion
      description: The trusted identity adapter supplies the verified actor. Authorization and Audit remain in Application.
      responses:
        '200':
          $ref: '#/components/responses/WithdrawResult'
        '401':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
      requestBody:
        $ref: '#/components/requestBodies/WithdrawVersion'
...
components:
  schemas:
    Uuid:
      $ref: ./schemas/common/identifiers.yaml#/$defs/Uuid
    Cursor:
      $ref: ./schemas/common/identifiers.yaml#/$defs/Cursor
    FieldError:
      $ref: ./schemas/errors/field-error.yaml
    Problem:
      type: object
      additionalProperties: false
      required:
      - type
      - title
      - status
      - code
      - traceId
      - retryable
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
          description: Stable machine code from spec/errors/error-registry.yaml.
        traceId:
          type: string
        retryable:
          type: boolean
        errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
        recovery:
          $ref: '#/components/schemas/CreateRecovery'
        exactRetry:
          type: boolean
          description: Only the same operation ID and payload may be retried.
    CreateRecovery:
      type: object
      additionalProperties: false
      required:
      - documentId
      - documentVersionId
      - fileId
      - recoveryEndpoint
      properties:
        documentId:
          $ref: '#/components/schemas/Uuid'
        documentVersionId:
          $ref: '#/components/schemas/Uuid'
        fileId:
          $ref: '#/components/schemas/Uuid'
        recoveryEndpoint:
          type: string
          format: uri-reference
...
    Version:
      $ref: ./schemas/document/models.yaml#/$defs/Version
    VersionList:
      $ref: ./schemas/document/models.yaml#/$defs/VersionList
...
    VersionMutationResult:
      $ref: ./schemas/document/models.yaml#/$defs/VersionMutationResult
    PublishResult:
      $ref: ./schemas/document/models.yaml#/$defs/PublishResult
    WithdrawResult:
      $ref: ./schemas/document/models.yaml#/$defs/WithdrawResult
...
    VersionWrite:
      $ref: ./schemas/document/commands.yaml#/$defs/VersionWrite
    RebaseVersion:
      $ref: ./schemas/document/commands.yaml#/$defs/RebaseVersion
    PublishVersion:
      $ref: ./schemas/document/commands.yaml#/$defs/PublishVersion
    WithdrawVersion:
      $ref: ./schemas/document/commands.yaml#/$defs/WithdrawVersion
...
  parameters:
    documentId:
      name: documentId
      in: path
      required: true
      description: Document identifier.
      schema:
        $ref: '#/components/schemas/Uuid'
    versionId:
      name: versionId
      in: path
      required: true
      description: Version identifier.
      schema:
        $ref: '#/components/schemas/Uuid'
...
    purpose:
      name: purpose
      in: query
      required: true
      description: Authorization purpose; it cannot be omitted to broaden access.
      schema:
        type: string
        enum:
        - published
        - authoring
        - history
...
    pageSize:
      name: pageSize
      in: query
      required: false
      description: Page size.
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    cursor:
      name: cursor
      in: query
      required: false
      description: Opaque, principal and query bound cursor.
      schema:
        $ref: '#/components/schemas/Cursor'
...
  requestBodies:
    RebaseVersion:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RebaseVersion'
    PublishVersion:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PublishVersion'
    WithdrawVersion:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/WithdrawVersion'
...
    VersionMultipart:
      required: true
      content:
        multipart/form-data:
          schema:
            type: object
            additionalProperties: false
            required:
            - request
            - files
            properties:
              request:
                $ref: '#/components/schemas/VersionWrite'
              files:
                type: array
                minItems: 1
                items:
                  type: string
                  format: binary
          encoding:
            request:
              contentType: application/json
            files:
              contentType: application/octet-stream
              headers:
                X-Part-Id:
                  description: Matches one VersionContentItem.partId; part order is immaterial.
                  schema:
                    type: string
  responses:
...
    Version:
      description: Version
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Version'
      headers:
        Cache-Control:
          description: Sensitive response is never shared or stored.
          schema:
            type: string
            const: private, no-store
    VersionList:
      description: VersionList
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/VersionList'
      headers:
        Cache-Control:
          description: Sensitive response is never shared or stored.
          schema:
            type: string
            const: private, no-store
...
    VersionMutationResult:
      description: VersionMutationResult
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/VersionMutationResult'
      headers:
        Cache-Control:
          description: Sensitive response is never shared or stored.
          schema:
            type: string
            const: private, no-store
    PublishResult:
      description: PublishResult
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PublishResult'
      headers:
        Cache-Control:
          description: Sensitive response is never shared or stored.
          schema: {type: string, const: 'private, no-store'}
    WithdrawResult:
      description: WithdrawResult
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/WithdrawResult'
      headers:
        Cache-Control:
          description: Sensitive response is never shared or stored.
          schema: {type: string, const: 'private, no-store'}
...
    Problem:
      description: RFC 9457 problem with stable machine code.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: about:blank
            title: Validation failed
            status: 422
            code: VALIDATION_FAILED
            traceId: example-trace
            retryable: false
            errors: []
      headers:
        Cache-Control:
          schema:
            type: string
            const: private, no-store